requirements.txt
aiogram
motor
mongo
python-dotenv
aiohttp
pytest
pytest-asyncio
black
isort
flake8